requirements.txt
asgiref==3.7.2
Django==4.2.6
django-filter==23.3
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-api-tracking==1.8.4
Markdown==3.5
psycopg2==2.9.9
PyJWT==2.8.0
python-environ==0.4.54
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
